Fixing Not Responding Results
when calling the function for first time it was pop up the prompt window but after hit enter it does not respond and showing the answers for that input, but that was because I was trying to call the function within the prompt inside it for it's Initial value so it takes an input yes but for the rest of the conditions it does not work because it was happening at the same time that takes input it "NO input yet".
so to fix that I made a function that return the value from prompt window and call that function when calling the Original function as a parameter.